// src/types.ts
export interface ApiKeys {
  googleClientId: string;
  firebaseApiKey: string;
  firebaseAuthDomain: string;
  firebaseProjectId: string;
  firebaseStorageBucket: string;
  firebaseMessagingSenderId: string;
  firebaseAppId: string;
  razorpayKeyId?: string;
  razorpayKeySecret?: string;
  paytmMid?: string;
  paytmMerchantKey?: string;
}